Refactor header to be one piece
