+++ src/css/_includes/global-header.scss
@@ -11,8 +11,11 @@
 }
 
 .global-header {
+  position: fixed;
   height: $headerHeight;
   width: 100%;
+  display: flex;
+  z-index: 2;
 
   .navbar-brand {
     display: flex;
@@ -23,6 +26,10 @@
     font-size: 0.5rem;
     padding-bottom: 0.7em;
     padding-top: 0.7em;
+
+    @media only screen and (min-width: 768px) {
+      width: 18rem;
+    }
 
     svg {
       display: block;
