+++ src/css/_includes/grid.scss
@@ -13,12 +13,6 @@
 .content-max {
   box-sizing: content-box;
   max-width: 1000px;
-  &.prose {
-
-    @media only screen and (min-width: 768px) {
-      margin-top: 4rem;
-    }
-}
 }
 
 
@@ -41,6 +35,7 @@ body {
 
     @media only screen and (min-width: 768px) {
       margin-left: 18rem;
+      padding-top: $headerHeight;
     }
   }
 }
@@ -51,7 +46,7 @@ body {
     max-height: 100%;
   }
 
-  .sidebar {
+  .sidebar, .brand {
     flex-basis: 270px;
     flex-shrink: 0;
   }
