+++ src/css/_includes/sidebar.scss
@@ -1,13 +1,15 @@
 .sidebar {
-
-  top: 0;
+  top: $headerHeight;
   bottom:0;
   flex-direction: column;
   display: flex;
   width: 100%;
   z-index: 1;
+  padding-top: $headerHeight;
+
   @media only screen and (min-width: 768px) {
     position: fixed;
+    padding-top: 0;
     width: 18rem;
   }
